No use overwriting install-sh when it already exists
git-svn-id: svn+ssh://svn.code.sf.net/p/firebird/code/firebird/branches/B2_5_Release@58977 65644016-39b1-43b1-bf79-96bc8fe82c15

+++ autogen.sh
@@ -71,7 +71,7 @@ $AUTORECONF --install --force --verbose || exit 1
 # automake 1.10 may show '1.10.x' as version while the directory is still named
 # automake-1.10
 am_ver=`automake --version|sed 's/.\+ //; s/\.[^.]\+$//; q'`
-cp /usr/share/automake-$am_ver*/install-sh $CONFIG_AUX_DIR || exit 1
+[ -f $CONFIG_AUX_DIR/install-sh ] || cp /usr/share/automake-$am_ver*/install-sh $CONFIG_AUX_DIR || exit 1
 
 # Hack to bypass bug in autoreconf - --install switch not passed to libtoolize,
 # therefore missing config.sub and confg.guess files
